UNTESTED LED refactor
This commit changes the LED pattern model so that each pattern is responsible for its own loop and position. This has several benefits, including allowing patterns that have more steps than there are LEDs in the setup, as well as patterns that want to wait for less than the default period. It also allows some optimizations around patterns that are not pixel-based (i.e. ring-based instead).